test: split workflow evidence smoke scenario
